#19 Moved the logic of the tree view which displays torrents to a separate class - TreeViewLogic
